chore: update all dependencies to latest versions
Bump 10 crates to their latest versions and fix breaking API changes:
- rusqlite 0.32→0.40, zip 2→8, reqwest 0.12→0.13, bollard 0.20→0.21,
  sevenz-rust2 0.12→0.21, rand 0.9→0.10, rpassword 5→7, toml 0.8→1,
  indicatif 0.17→0.18, actix-web-lab 0.24→0.26

+++ src/db/mods.rs
@@ -118,9 +118,9 @@ impl Database {
         )?;
         let rows = stmt.query_map([], |row| {
             let m = row_to_installed_mod(row)?;
-            let count: usize = row.get(8)?;
+            let count: i64 = row.get(8)?;
             let size: i64 = row.get(9)?;
-            Ok((m, count, size))
+            Ok((m, count as usize, size))
         })?;
         rows.collect()
     }
